stubbing out alarm module.
