+++ include/alarm.h
@@ -1,0 +1,32 @@
+/**
+ * @file alarm.h
+ * @brief TODO.
+ *
+ */
+
+
+
+
+#ifndef ALARM_H
+#define	ALARM_H
+
+
+
+
+#include "gui.h"
+
+
+
+
+//
+//void calendar_set_default_configuration( gui_calendar_s * const calendar );
+//
+//
+////
+//void calendar_render(
+//        gui_s * const gui,
+//        gui_calendar_s * const calendar );
+
+
+
+#endif	/* ALARM_H */
