+++ include/calendar.h
@@ -14,7 +14,6 @@
 
 
 #include "gui.h"
-#include "display_manager.h"
