+++ include/font.h
@@ -1,0 +1,27 @@
+/**
+ * @file font.h
+ * @brief TODO.
+ *
+ */
+
+
+
+
+#ifndef FONT_H
+#define FONT_H
+
+
+
+
+#include "gui.h"
+
+
+
+
+//
+void *font_get( const gui_text_font_kind font_kind );
+
+
+
+
+#endif /* FONT_H */
