+++ src/alarm.c
@@ -1,0 +1,59 @@
+/**
+ * @file alarm.c
+ * @brief TODO.
+ *
+ */
+
+
+
+
+#include <stdlib.h>
+#include <stdio.h>
+#include <string.h>
+#include <unistd.h>
+#include <time.h>
+
+#include <VG/openvg.h>
+#include <VG/vgu.h>
+#include <fontinfo.h>
+#include <shapes.h>
+
+#include "gui.h"
+#include "time_domain.h"
+#include "font.h"
+#include "alarm.h"
+
+
+
+
+// *****************************************************
+// static global types/macros
+// *****************************************************
+
+
+
+
+// *****************************************************
+// static global data
+// *****************************************************
+
+
+
+
+// *****************************************************
+// static declarations
+// *****************************************************
+
+
+
+
+// *****************************************************
+// static definitions
+// *****************************************************
+
+
+
+
+// *****************************************************
+// public definitions
+// *****************************************************
